Start working on finger detection.
This doesn't work yet.  The sync loop never sees a successful
return from the hardware, likely due to the 3 different bits
in the two different replies to previous messages.  The async
loop looks like it deadlocks itself, since the second call to
start_finger_detection() hangs in the send() function after sending
the second poll to the hardware.

Signed-off-by: Ray Lehtiniemi <rayl@mail.com>